Update TreeAdapter.js
Use provided `leafField` property name instead of the hard-coded `$leaf`.

Note that processList method in the TreeAdapter does `result.push(...nonLeafs, ...leafs);` which basically always returns non leaf elements before leaf elements. The sorting logic that we added to TreeAdapter sort method does not have any affect on the order of the elements, as the sorting logic is applied before the processing is done. :)

We should probably consider `foldersFirst` property when processing the results, like it's done in the sort method.

Last, can we use preSorters to achieve this sorting logic, when foldersFirst is set to true.